feat(wxapi): 新增视频号小店上传图片接口 feat(wxapi): 新增视频号橱窗上下架商品接口 feat(wxapi): 标记微信小店相关接口及模型为已废弃 feat(wxapi): 新增视频号小店商品类目相关接口 feat(wxapi): 新增视频号小店物流相关接口 feat(wxapi): 新增视频号小店运费模板相关接口 feat(wxapi): 新增视频号小店地址管理相关接口 feat(wxapi): 新增视频号小店商品操作相关接口 feat(wxapi): 新增视频号小店商品限时抢购相关接口 feat(wxapi): 调整部分微信商品库相关接口模型的命名方式 feat(wxapi): 新增视频号小店订单相关接口 feat(wxapi): 新增视频号小店售后单相关接口 feat(wxapi): 新增视频号小店优惠券相关接口 docs: 完善文档 feat(wxapi): 调整腾讯云服务 TCB 相关接口及模型的命名方式 feat(wxapi): 新增微信云托管服务管理相关接口 提交者： Fu Diwei chore(wxapi): 文档注释

+++ src/DH.SKIT.FlurlHttpClient.Wechat.Api/Models/UnionPromoter/Product/UnionPromoterProductListResponse.cs
@@ -73,7 +73,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Api.Models
                                         /// </summary>
                                         [Newtonsoft.Json.JsonProperty("templateId")]
                                         [System.Text.Json.Serialization.JsonPropertyName("templateId")]
-                                        public int TemplateId { get; set; }
+                                        public long TemplateId { get; set; }
                                     }
                                 }
 
@@ -512,7 +512,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Api.Models
                         [Newtonsoft.Json.JsonProperty("couponId")]
                         [System.Text.Json.Serialization.JsonPropertyName("couponId")]
                         [System.Text.Json.Serialization.JsonNumberHandling(System.Text.Json.Serialization.JsonNumberHandling.AllowReadingFromString)]
-                        public int? CouponId { get; set; }
+                        public long? CouponId { get; set; }
 
                         /// <summary>
                         /// 获取或设置优惠券详细信息。
